fix helper bug, move rollup configs into one file
